Split notif structs into diff files
