--- Guardian/Notification.swift
@@ -1,386 +1,0 @@
-// Notification.swift
-//
-// Copyright (c) 2016 Auth0 (http://auth0.com)
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in
-// all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-// THE SOFTWARE.
-
-import Foundation
-
-/**
- Stores the default Guardian Push notification category and actions identifiers.
- With these values you can create the `UNNotificationCategory` and `UNNotificationAction` to register
- when notification access is requested using the UserNotification.framework.
-
- ```
- let guardianCategory = Guardian.AuthenticationCategory.default
-
- // Set up guardian notifications actions
- let acceptAction = UNNotificationAction(
-    identifier: guardianCategory.allow.identifier,
-    title: NSLocalizedString("Allow", comment: "Accept Guardian authentication request"),
-    options: [.authenticationRequired] // Always request local AuthN
- )
- let rejectAction = UNNotificationAction(
-    identifier: guardianCategory.reject.identifier,
-    title: NSLocalizedString("Deny", comment: "Reject Guardian authentication request"),
-    options: [.destructive, .authenticationRequired] // Always request local AuthN
- )
-
- // Set up guardian notification category
- let category = UNNotificationCategory(
-    identifier: guardianCategory.identifier,
-    actions: [acceptAction, rejectAction],
-    intentIdentifiers: [],
-    options: [])
-
- // Request for AuthZ
- UNUserNotificationCenter.current().requestAuthorization(options: [.badge, .sound]) { granted, error in
-    guard granted else {
-        return print("Permission not granted")
-    }
-    if let error = error {
-        return print("failed with error \(error)")
-    }
-
-    // Register guardian notification category
-    UNUserNotificationCenter.current().setNotificationCategories([category])
-
-    // Check AuthZ status to trigger remote notification registration
-    UNUserNotificationCenter.current().getNotificationSettings() { settings in
-        guard settings.authorizationStatus == .authorized else {
-            return print("not authorized to use notifications")
-        }
-        DispatchQueue.main.async { application.registerForRemoteNotifications() }
-    }
- }
- ```
-
- Then in your User notification delegate, handle the notification action using the identifier provided by iOS
-
- ```
- func userNotificationCenter(_ center: UNUserNotificationCenter, didReceive response: UNNotificationResponse, withCompletionHandler completionHandler: @escaping () -> Void) {
-    let identifier = response.actionIdentifier
-    let userInfo = response.notification.request.content.userInfo
-    if let notification = Guardian.notification(from: userInfo) {
-        /* Get the enrollment that matches the notification ... */
-        let enrollment: Enrollment = ...
-        Guardian
-        .authentication(forDomain: "tenant.guardian.auth0.com", andEnrollment: enrollment)
-        .handleAction(withIdentifier: identifier, notification: notification)
-        .start { result in
-            completionHandler()
-        }
-    } else {
-        // Other type of notification
-        completionHandler()
-    }
- }
- ```
-
- */
-public struct AuthenticationCategory {
-
-    public struct Action: Equatable {
-        public let identifier: String
-    }
-
-    public let identifier: String
-    public let allow: Action
-    public let reject: Action
-
-    private init(identifier: String) {
-        self.identifier = identifier
-        self.allow = Action(identifier: "\(self.identifier).accept")
-        self.reject = Action(identifier: "\(self.identifier).reject")
-    }
-
-    private static let defaultIdentifier = "com.auth0.notification.authentication"
-
-    public static var `default`: AuthenticationCategory {
-        return AuthenticationCategory(identifier: defaultIdentifier)
-    }
-}
-
-/**
- A Guardian Notification contains data about an authentication request.
- 
- You get one of this from the push notification data, for example in your 
- `AppDelegate` you should have something like this:
-
- ```
- func application(application: UIApplication, didReceiveRemoteNotification userInfo: [AnyHashable: Any]) {
-    if let notification = Guardian.notification(from: userInfo) {
-        // the push notification is a Guardian authentication request
-        // do something with it
-        print(notification)
-    }
- }
- ```
- */
-public protocol Notification {
-
-    /**
-     The Guardian server that sent the notification
-     */
-    var domain: String { get }
-
-    /**
-     The id of the Guardian `Enrollment`
-     */
-    var enrollmentId: String { get }
-
-    /**
-     The transaction token, used to identify the authentication request
-     */
-    var transactionToken: String { get }
-
-    /**
-     The challenge sent by the server. The same challenge, signed, should be 
-     sent back when trying to allow or reject an authentication request
-     */
-    var challenge: String { get }
-
-    /**
-     The source (Browser & OS) where the authentication request was initiated,
-     if available
-     */
-    var source: Source? { get }
-
-    /**
-     The location where the request was initiated, if available
-     */
-    var location: Location? { get }
-
-    /**
-     The date/time when the authentication request was initiated
-     */
-    var startedAt: Date { get }
-}
-
-/**
- The source (Browser & OS) of an authentication request
- */
-public protocol Source {
-
-    /**
-     The operating system data, if available
-     */
-    var os: OS? { get }
-
-    /**
-     The browser data, if available
-     */
-    var browser: Browser? { get }
-}
-
-/**
- The browser data of an authentication request
- */
-public protocol Browser {
-
-    /**
-     The name of the browser
-     */
-    var name: String { get }
-
-    /**
-     The version of the browser, if available
-     */
-    var version: String? { get }
-}
-
-/**
- The OS data of an authentication request
- */
-public protocol OS {
-
-    /**
-     The name of the operating system
-     */
-    var name: String { get }
-
-    /**
-     The version of the operating system, if available
-     */
-    var version: String? { get }
-}
-
-/**
- The geographical location of an authentication request
- */
-public protocol Location {
-
-    /**
-     The name of the (approximate) location, if available
-     */
-    var name: String? { get }
-
-    /**
-     The approximate latitude, if available
-     */
-    var latitude: Double? { get }
-
-    /**
-     The approximate longitude, if available
-     */
-    var longitude: Double? { get }
-}
-
-struct AuthenticationNotification: Notification, CustomDebugStringConvertible, CustomStringConvertible {
-
-    let domain: String
-    let enrollmentId: String
-    let transactionToken: String
-    let challenge: String
-    let source: Source?
-    let location: Location?
-    let startedAt: Date
-
-    init(domain: String, enrollmentId: String, transactionToken: String, challenge: String, startedAt: Date, source: Source?, location: Location?) {
-        self.domain = domain
-        self.enrollmentId = enrollmentId
-        self.transactionToken = transactionToken
-        self.challenge = challenge
-        self.source = source
-        self.location = location
-        self.startedAt = startedAt
-    }
-
-    init?(userInfo: [AnyHashable: Any]) {
-        guard
-            let json = userInfo as? [String: Any],
-            let aps = json["aps"] as? [String: Any],
-            let category = aps["category"] as? String, category == AuthenticationCategory.default.identifier
-            else { return nil }
-        let locale = Locale(identifier: "en_US_POSIX")
-        let formatter = DateFormatter()
-        formatter.locale = locale
-        formatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSSZ"
-        guard
-            let mfa = json["mfa"] as? [String: Any],
-            let enrollmentId = mfa["dai"] as? String,
-            let token = mfa["txtkn"] as? String,
-            let when = mfa["d"] as? String,
-            let startedAt = formatter.date(from: when),
-            let domain = mfa["sh"] as? String,
-            let challenge = mfa["c"] as? String
-            else { return nil }
-        let source = AuthenticationSource(fromJSON: mfa["s"])
-        let location = AuthenticationLocation(fromJSON: mfa["l"])
-
-        self.init(domain: domain, enrollmentId: enrollmentId, transactionToken: token, challenge: challenge, startedAt: startedAt, source: source, location: location)
-    }
-
-    var description: String {
-        return "enrollmentId: <\(self.enrollmentId)> txToken: <\(self.transactionToken)> challenge: <\(self.challenge)> startedAt: \(self.startedAt)"
-    }
-
-    var debugDescription: String {
-        return "domain: <\(self.domain)> enrollmentId: <\(self.enrollmentId)> txToken: <\(self.transactionToken)> challenge: <\(self.challenge)> source: <\(String(describing: self.source))> location: <\(String(describing: self.location))> startedAt: \(self.startedAt)"
-    }
-}
-
-struct AuthenticationSource: Source, CustomDebugStringConvertible, CustomStringConvertible {
-
-    class NamedSource: NSObject, OS, Browser {
-        let name: String
-        let version: String?
-
-        init(name: String, version: String?) {
-            self.name = name
-            self.version = version
-        }
-    }
-
-    let os: OS?
-    let browser: Browser?
-
-    init?(fromJSON json: Any?) {
-        guard let source = json as? [String: Any] else {
-            return nil
-        }
-
-        let browser: Browser?
-        let os: OS?
-        if let data = source["b"] as? [String: Any], let name = data["n"] as? String {
-            let version = data["v"] as? String
-            browser = NamedSource(name: name, version: version)
-        } else {
-            browser = nil
-        }
-        if let data = source["os"] as? [String: Any], let name = data["n"] as? String {
-            let version = data["v"] as? String
-            os = NamedSource(name: name, version: version)
-        } else {
-            os = nil
-        }
-
-        if os == nil && browser == nil {
-            return nil
-        }
-        self.os = os
-        self.browser = browser
-    }
-
-    var description: String {
-        let osName = self.os?.name ?? "Unknown OS"
-        let osVersion = self.os?.version != nil ? "(\(String(describing: self.os?.version))" : ""
-        let browserName = self.browser?.name ?? "Unknown Browser"
-        let browserVersion = self.browser?.version != nil ? "(\(String(describing: self.browser?.version))" : ""
-        return "\(osName) \(osVersion)".trimmingCharacters(in: .whitespaces) + " \(browserName) \(browserVersion)".trimmingCharacters(in: .whitespaces)
-    }
-
-    var debugDescription: String {
-        return self.description
-    }
-}
-
-struct AuthenticationLocation: Location, CustomDebugStringConvertible, CustomStringConvertible {
-
-    let name: String?
-    let latitude: Double?
-    let longitude: Double?
-
-    init?(fromJSON json: Any?) {
-        guard let location = json as? [String: Any] else {
-            return nil
-        }
-
-        name = location["n"] as? String
-        let latitudeValue = location["lat"]
-        let longitudeValue = location["long"]
-        if let latitudeString = latitudeValue as? String,
-            let longitudeString = longitudeValue as? String {
-            latitude = Double(latitudeString)
-            longitude = Double(longitudeString)
-        } else {
-            latitude = (latitudeValue as? NSNumber)?.doubleValue
-            longitude = (longitudeValue as? NSNumber)?.doubleValue
-        }
-    }
-
-    var description: String {
-        return "\(self.name ?? "Unknown") (\(String(describing: self.latitude)), \(String(describing: self.longitude)))"
-    }
-
-    var debugDescription: String {
-        return self.description
-    }
-
-}
