+++ Guardian/Notification/AuthenticationLocation.swift
@@ -1,0 +1,57 @@
+// AuthenticationLocation.swift
+//
+// Copyright (c) 2018 Auth0 (http://auth0.com)
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+import Foundation
+
+struct AuthenticationLocation: Location, CustomDebugStringConvertible, CustomStringConvertible {
+
+    let name: String?
+    let latitude: Double?
+    let longitude: Double?
+
+    init?(fromJSON json: Any?) {
+        guard let location = json as? [String: Any] else {
+            return nil
+        }
+
+        name = location["n"] as? String
+        let latitudeValue = location["lat"]
+        let longitudeValue = location["long"]
+        if let latitudeString = latitudeValue as? String,
+            let longitudeString = longitudeValue as? String {
+            latitude = Double(latitudeString)
+            longitude = Double(longitudeString)
+        } else {
+            latitude = (latitudeValue as? NSNumber)?.doubleValue
+            longitude = (longitudeValue as? NSNumber)?.doubleValue
+        }
+    }
+
+    var description: String {
+        return "\(self.name ?? "Unknown") (\(String(describing: self.latitude)), \(String(describing: self.longitude)))"
+    }
+
+    var debugDescription: String {
+        return self.description
+    }
+
+}
